client ne crash pas avec emoji

+++ srcs/client.c
@@ -6,7 +6,7 @@
 /*   By: myanez-p <myanez-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/18 11:48:45 by myanez-p          #+#    #+#             */
-/*   Updated: 2023/04/24 11:09:33 by myanez-p         ###   ########.fr       */
+/*   Updated: 2023/09/04 17:06:05 by myanez-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,23 +25,26 @@ void	send_message(char *message, int pid_server)
 	i = 0;
 	while (message[i])
 	{
-		tmp = message[i];
-		while (message[i] > 0)
+		if (message[i] >= 0 && message[i] <= 127)
 		{
-			if (message[i] % 2 == 0)
+			tmp = message[i];
+			while (message[i] > 0)
+			{
+				if (message[i] % 2 == 0)
+					kill(pid_server, SIGUSR1);
+				else
+					kill(pid_server, SIGUSR2);
+				message[i] /= 2;
+				usleep(100);
+			}
+			while (tmp < 128)
+			{
 				kill(pid_server, SIGUSR1);
-			else
-				kill(pid_server, SIGUSR2);
-			message[i] /= 2;
-			usleep(100);
+				usleep(100);
+				tmp *= 2;
+			}
 		}
-		while (tmp < 128)
-		{
-			kill(pid_server, SIGUSR1);
-			usleep(100);
-			tmp *= 2;
-		}
-		i ++;
+		i++;
 	}
 }
